Eyra: a new way to build Rust programs entirely written in Rust.
Eyra is similar to Mustang, except that instead of a custom target,
it just needs to link with `-nostartfiles`. This means it doesn't need
`-Z build-std`; it can slide in underneath normal std builds. See the
example in example-crates/eyra-example for details.

To support this, give c-scape an option to implement its `malloc` in
terms of rustix-dlmalloc directly rather than using the global allocator,
allocator, since the global allocator may be implemented using
`malloc`.

And, give c-scape a "define-mem-functions" option to disable definitions
of the C memory functions such as `memcpy`. It's not clear exactly when
Rust links in the compiler_builtins from the sysroot and when it uses
the one from the cargo dependency, and the sysroot version doesn't
appear to have these functions defined, so just let users decide which
one they think they're going to get.