Stop having Eyra try to be a drop-in replacement for libc.
For users already using `libc`, it's inconvenient to add Eyra as a
`libc` dependency, and for users not using `libc`, having Eyra be a
passthrough isn't useful.

So instead, change the main instructions to use `extern crate eyra`
instead of `extern crate libc`, and remove Eyra's `use libc::*;`.
This makes it simpler overall, and easier to add Eyra to an existing
codebase.

Also, add an example of adding Eyra as an optional dependency.