Implement more features.
Implement `SO_PEERCRED`, a few more `sysconf` codes, and `accept` taking
a null destination address or getting an `AF_UNSPEC` address.

And add a special case for Unix-domain addresses where the caller has
forgotten to add 1 for the NUL terminator to fix it up, so that rustix
doesn't have to be aware of this.